signer: move GPGSigner to _gpg_signer.py
