+++ securesystemslib/signer/__init__.py
@@ -5,12 +5,12 @@ This module provides extensible interfaces for public keys and signers:
 Some implementations are provided by default but more can be added by users.
 """
 from securesystemslib.signer._gcp_signer import GCPSigner
+from securesystemslib.signer._gpg_signer import GPGSigner
 from securesystemslib.signer._hsm_signer import HSMSigner
 from securesystemslib.signer._key import KEY_FOR_TYPE_AND_SCHEME, Key, SSlibKey
 from securesystemslib.signer._signature import GPGSignature, Signature
 from securesystemslib.signer._signer import (
     SIGNER_FOR_URI_SCHEME,
-    GPGSigner,
     SecretsHandler,
     Signer,
     SSlibSigner,
