+++ securesystemslib/signer/_signer.py
@@ -6,10 +6,9 @@ from abc import ABCMeta, abstractmethod
 from typing import Callable, Dict, Optional, Type
 from urllib import parse
 
-import securesystemslib.gpg.functions as gpg
 import securesystemslib.keys as sslib_keys
 from securesystemslib.signer._key import Key, SSlibKey
-from securesystemslib.signer._signature import GPGSignature, Signature
+from securesystemslib.signer._signature import Signature
 
 logger = logging.getLogger(__name__)
 
@@ -212,62 +211,3 @@ class SSlibSigner(Signer):
         """
         sig_dict = sslib_keys.create_signature(self.key_dict, payload)
         return Signature(**sig_dict)
-
-
-class GPGSigner(Signer):
-    """A securesystemslib gpg implementation of the "Signer" interface.
-
-    Provides a sign method to generate a cryptographic signature with gpg, using
-    an RSA, DSA or EdDSA private key identified by the keyid on the instance.
-    """
-
-    def __init__(
-        self, keyid: Optional[str] = None, homedir: Optional[str] = None
-    ):
-        self.keyid = keyid
-        self.homedir = homedir
-
-    @classmethod
-    def from_priv_key_uri(
-        cls,
-        priv_key_uri: str,
-        public_key: Key,
-        secrets_handler: Optional[SecretsHandler] = None,
-    ) -> "GPGSigner":
-        raise NotImplementedError("Incompatible with private key URIs")
-
-    def sign(self, payload: bytes) -> GPGSignature:
-        """Signs a given payload by the key assigned to the GPGSigner instance.
-
-        Calls the gpg command line utility to sign the passed content with the
-        key identified by the passed keyid from the gpg keyring at the passed
-        homedir.
-
-        The executed base command is defined in
-        securesystemslib.gpg.constants.gpg_sign_command.
-
-        Arguments:
-            payload: The bytes to be signed.
-
-        Raises:
-            securesystemslib.exceptions.FormatError:
-                If the keyid was passed and does not match
-                securesystemslib.formats.KEYID_SCHEMA.
-
-            ValueError: the gpg command failed to create a valid signature.
-            OSError: the gpg command is not present or non-executable.
-            securesystemslib.exceptions.UnsupportedLibraryError: the gpg
-                command is not available, or the cryptography library is
-                not installed.
-            securesystemslib.gpg.exceptions.CommandError: the gpg command
-                returned a non-zero exit code.
-            securesystemslib.gpg.exceptions.KeyNotFoundError: the used gpg
-                version is not fully supported and no public key can be found
-                for short keyid.
-
-        Returns:
-            Returns a "GPGSignature" class instance.
-        """
-
-        sig_dict = gpg.create_signature(payload, self.keyid, self.homedir)
-        return GPGSignature(**sig_dict)
